Fix FFI namespaces to match folders; move RingBuffer to Audio/
Follow-up cleanup to the feature-folder reorg (e8dc8e9), which renamed
folders but left namespaces and one misplaced file behind.

- Rename namespace LiveKit.Internal.FFIClients* -> LiveKit.Internal.FFI*
  so it matches the Internal/FFI/ folder (the reorg renamed FFIClients/
  -> FFI/ but kept the old namespace). Updated all usings across Runtime
  and Tests.
- Collapse the dead .Pools.ObjectPool namespace into .Pools (the
  ObjectPool/ subfolder was dissolved in the reorg) and drop the now
  redundant self-imports it left behind.
- Fix two malformed namespaces in files the reorg moved: SmartWrap.cs
  had an IDE-autogenerated path-as-namespace, and Pools.cs had a trailing
  "//" in its namespace declaration.
- Move RingBuffer.cs into Audio/ (its only consumer is AudioStream.cs),
  alongside the other audio helpers pulled in by the reorg.

No behavioural change. Verified by compiling the LiveKit assembly (csc)
from the Meet sample's generated rsp.
