+++ Runtime/Scripts/Audio/AudioResampler.cs
@@ -1,5 +1,5 @@
 using System;
-using LiveKit.Internal.FFIClients.Requests;
+using LiveKit.Internal.FFI.Requests;
 using LiveKit.Internal;
 using LiveKit.Proto;
