+++ Runtime/Scripts/Audio/AudioStream.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 using LiveKit.Internal;
 using LiveKit.Proto;
 using System.Runtime.InteropServices;
-using LiveKit.Internal.FFIClients.Requests;
+using LiveKit.Internal.FFI.Requests;
 
 namespace LiveKit
 {
