+++ Runtime/Scripts/Core/Room.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using LiveKit.Internal;
 using LiveKit.Proto;
 using System.Runtime.InteropServices;
-using LiveKit.Internal.FFIClients.Requests;
+using LiveKit.Internal.FFI.Requests;
 using UnityEngine;
 
 namespace LiveKit
