+++ Runtime/Scripts/Core/TrackPublication.cs
@@ -1,5 +1,5 @@
 using LiveKit.Internal;
-using LiveKit.Internal.FFIClients.Requests;
+using LiveKit.Internal.FFI.Requests;
 using LiveKit.Proto;
 
 namespace LiveKit
