+++ Runtime/Scripts/DataStreams/TextDataStream.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Collections.Generic;
-using LiveKit.Internal.FFIClients.Requests;
+using LiveKit.Internal.FFI.Requests;
 using LiveKit.Internal;
 using LiveKit.Proto;
