+++ Runtime/Scripts/Internal/FFI/Pools/Memory/IMemoryPool.cs
@@ -1,6 +1,6 @@
 using Google.Protobuf;
 
-namespace LiveKit.Internal.FFIClients.Pools.Memory
+namespace LiveKit.Internal.FFI.Pools.Memory
 {
     public interface IMemoryPool
     {
