+++ Runtime/Scripts/Internal/FFI/Pools/Pools.cs
@@ -1,9 +1,8 @@
 using System;
-using LiveKit.Internal.FFIClients.Pools.ObjectPool;
 using LiveKit.Proto;
 using UnityEngine.Pool;
 
-namespace LiveKit.Internal.FFIClients.Pools//
+namespace LiveKit.Internal.FFI.Pools
 {
     public static class Pools
     {
