+++ Runtime/Scripts/Video/RtcVideoSource.cs
@@ -5,7 +5,7 @@ using LiveKit.Internal;
 using UnityEngine.Rendering;
 using Unity.Collections;
 using Unity.Collections.LowLevel.Unsafe;
-using LiveKit.Internal.FFIClients.Requests;
+using LiveKit.Internal.FFI.Requests;
 using System.Collections;
 
 namespace LiveKit
